Added performance tests for key generation
